+++ HISTORY.rst
@@ -7,3 +7,19 @@ History
 ++++++++++++++++++
 
 * First release on PyPI.
+
+
+0.1.3 (2016-04-22)
+++++++++++++++++++
+
+* Added ``template`` setting
+* Added ``get_email`` setting
+* Added ``get_language`` setting
+* Added ``get_template`` setting
+* Added ``get_fail_backlink`` setting
+* Added ``get_backlink_url`` setting
+* Added ``auto_capture`` setting
+* Updated test geteway url to https://testpay.kkb.kz
+
+
+